// frontend/src/app/models/izvestaj.ts
export class Izvestaj{
    pacijent: string = '';
    lekarKorIme: string = '';
    lekarIme: string = '';
    lekarPrezime: string = '';
    lekarSpecijalizacija: string = '';
    datum: string = '';
    vreme: string = '';
    razlogDolaska: string = '';
    dijagnoza: string = '';
    preporucenaTerapija: string = '';
    preporucenDatumKontrole: string = '';
}